Refactors note creation for Twingle submissions
Adds note creation for contributions created as SEPA mandates.

Moves note creation logic from the Submit API to the Submission class.

This change centralizes the creation of contact and contribution notes
related to Twingle submissions, promoting code reuse and maintainability.

Improves the readability of the Submit API by reducing its complexity.

Fixes #96
